Official release for the webinar fixes (#2)
This updates the configuration provided for the webinar to v2.3.0 of the Publish2 app.

Note that the implementation of the farm submission code has changed a bit. The publish_file.py hook that comes with the publisher now supports the publish_user property on a publish item, which does not require our farm_wrapper.py to implement get_publisher_user(). We can now simply assign a value to it and we're good to go.